Initial pass at a beginning sketch for API server.
Simple implementations for login (JWT authentication token acquisition),
creating new user profiles, and obtaining profile for the current
authenticated user + for a profile by ID.

This is very much a starting point and there's a lot missing from
here. In no specific order, known issues include:
* A lack of unit tests
* Mock / stub data included in this project - this will migrate to
  a separate project and service for data interfacing.
* Lack of password hashing - this will be addressed when we have a
  database and data service to serve data to this API
* Fixed JWT signing string - needs to move to getting from environment
* Password field being propogated out to profiles returned by the API.
  Currently filtered during marshalling to be empty strings,
  likely this can be improved.
* Internal error handling can be improved.

The intent of this addition is to give a starting point for project participants
to be able to start working on the API and to give a concrete forum for
prototyping potential API additions.

This is the author's first Go project - pointers and tips on style,
conventions, ways to improve code or anything else are welcomed.