Add proposals for POST of user profile and GET of profile by ID
Adds NotFound common response to be used as return response
both for when a requested resource is not found, and when
a requested resource is not one the requesting user is permitted
to access. This common response is proposed so the API doesn't
leak information about resources being valid but not for the user
that is requesting them.