Update API: change from camelCase to snake_case naming
This changes API schema naming and field naming from the mix
of camelCase and PascalCase previously proposed, to consistent
snake_case naming.
The initial sketch code maintains PascalCase usage for field names
in the data models as Golang requires struct fields to begin with a capital letter
in order for the JSON package to see and export these.

As part of this, additional de-capitalization of certain words
in comments was performed.