Restore previous password marshalling behavior
Previous change to add an entry to the marshalling string to ignore
the 'password' field causes the field to be omitted when marshalling
the struct (converting to JSON) - which is good - but also causes it
to be ignored and discarded when unmarshalling the JSON (converting
JSON to a LasagnaLoveUser) - which causes Profile creation to fail.

Restored previous behavior, which is less than elegant but restores
ability to make new profiles.