Split Docker local configuration into debugger and non-debugger running
The Delve debugger for Golang halts program execution on startup,
meaning the previous configuration required a debugger to attach
in order for the server to run. Execution outside of a debugger
was not functioning.
Giving Delve the --continue=true flag at instantiation stops this
but then does not allow VS Code or debuggers to attach.

To address this, this PR splits the Docker configurations into two:
* One for running locally in/via a debugger (e.g. VS Code, JetBrains, etc.)
* A second for running locally outside of a debugger

This also renames Docker compose files and targets for clarity
and includes other minor touchups.