Start of patch implementation for /profile/[:userID]
Updates simple fields in Profile objects currently.
Crashes when given an embedded struct (such as Attestations).
Needs to handle passwords properly.

Tried using a data model for patching, but the JSON binding pulls
in empty/unset fields as well as fields for updating, which leads
to big problems. The method used in this commit may not be optimal.